do_actions: set the default action return code to 0
The return value passed to m_action::finish() is inserted into the
action table's status column. If an action doesn't have a specific
return set by the output of the job it ran, then 'OK' was being passed
through to the database. Consequently the query in m_action::finish()
always failed.